New numlab exercises filter.
Exercises Filter-Plugin.
This filter masquerades the response of an request to
/numlab/exercises/<id>. With a query string the request only contain
toplevel properties with simple string type. For now these toplevel
properties are:

  * postTime
  * identifier
  * comment
  * name
  * description
  * environment

The querystring parameter is known as: properties. You're allowed to
note multiple toplevel properties, spaced by comma. Please be patient
not to include any spaces in the list.

Working ressource notations are:

http://ecs.uni-stuttgart.de/numlab/exercises/27?properties=name
http://ecs.uni-stuttgart.de/numlab/exercises/27?properties=name,description
http://ecs.uni-stuttgart.de/numlab/exercises/27?properties=name,description,comment,identifier

Problematic/forbidden notations are:

http://ecs.uni-stuttgart.de/numlab/exercises/27?properties= name
http://ecs.uni-stuttgart.de/numlab/exercises/27?properties=name ,description
http://ecs.uni-stuttgart.de/numlab/exercises/27?properties =name,description,comment , identifier